Updated styling
Buttons now don't have that weird thing going on with the appearance they did last time. Updated so have a hover effect of just the border increasing, but no color changes (as intended). Removed a previous test text. Added keys to (as of now) both buttons, so they can be individually styled if needed (was needed for clear chat history).