Started docker stuff.
Created the docker file so that when we got to deliver the executable it will all be in one thing. Filled requirements.txt with the necessary requirements. Edited the readme with some further instructions for this. Note that more research will be needed as it looks like two images will need to be created (one for Bob, and one for Ollama), and they will need to be run simultaneously. This will be looked into further.